Handle error status from the Resto API
